pyproject.toml
[project]
name = "django-telegram-reminder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-select2>=8.4.3",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
]
